further work on fragmented extraction dialog
